Partially implement DJINI:PERSIST results of JEANNIE:READ-RDF in memory.
Currently just a hack of grabbing a list of DJINI:TRIPLE whose first
element contains the serialized string representation.  We're going to
have to clean up things afterwards anyways, so we might as well start
with the naieve approach.

Try it with a pathname containing valid RDF N3:

	(djini:persist "~/work/abcl/abcl.rdf")

which will slice 'n dice everything it can find into DJINI:*MODEL*.